token simulation

+++ app/Http/Controllers/API/UserAPI.php
@@ -1,16 +1,22 @@
 <?php
 
 namespace App\Http\Controllers\API;
+use App\Core\Token as Token;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
-use function PHPUnit\Framework\isEmpty;
+use \Illuminate\Http\JsonResponse;
 
 class UserAPI extends Controller
 {
-    function getAllUsers()
+    function getAllUsers(Request $request): JsonResponse
     {
+        $token = new Token();
+        if (!$token->checkTokenForAvailability($request->bearerToken()))
+            return response()->json(['message' => 'Unauthorized '], 401 );
+
+
         $result =  User::all();
         if ($result){
             return response()->json(['message' => 'the request has been succeeded','data' => $result],201 );
@@ -21,9 +27,14 @@ class UserAPI extends Controller
         }
     }
 
-    function getUsersByID($user_id)
+    function getUsersByID(Request $request, $userId): JsonResponse
     {
-        $result = User::find($user_id);
+        $token = new Token();
+        if (!$token->checkTokenForAvailability($request->bearerToken())){
+            return response()->json(['message' => 'Unauthorized '], 401 );
+        }
+
+        $result = User::find($userId);
         if ($result){
             return response()->json(['message' => 'the request has been succeeded','data' => $result],201 );
         }
@@ -31,20 +42,24 @@ class UserAPI extends Controller
         {
             return response()->json(['message' => 'Not found'], 404);
         }
-
-
     }
 
-    function createUser(Request $request )
+    function createUser(Request $request ): JsonResponse
     {
+        $token = new Token();
+        if (!$token->checkTokenForAvailability($request->bearerToken())){
+            return response()->json(['message' => 'Unauthorized '], 401 );
+        }
+
         $user = new User;
         $user->name = $request->name;
         $user->age = $request->age;
         $user->email = $request->email;
+        $user->password = md5($request->password);
         $result =  $user->save();
         if ($result)
         {
-            return response()->json(['message' => 'User has been created','data' => $user] , 201  );
+            return response()->json(['message' => 'User has been created'], 200);
         }
         else
         {
@@ -53,8 +68,13 @@ class UserAPI extends Controller
 
     }
 
-    function updateUserByID($user_id,Request $request)
+    function updateUserByID($user_id,Request $request): JsonResponse
     {
+        $token = new Token();
+        if (!$token->checkTokenForAvailability($request->bearerToken())){
+            return response()->json(['message' => 'Unauthorized '], 401 );
+        }
+
         $user = User::find($user_id);
         $user->name = $request->name;
         $user->age = $request->age;
@@ -63,7 +83,7 @@ class UserAPI extends Controller
 
         if ($result)
         {
-            return response()->json(['message' => 'User has been updated','data' => $user] , 201  );
+            return response()->json(['message' => 'User has been updated'], 200);
         }
         else
         {
@@ -72,13 +92,17 @@ class UserAPI extends Controller
 
     }
 
-    function deleteUserByID($user_id)
+    function deleteUserByID(Request $request, $user_id): JsonResponse
     {
+        $token = new Token();
+        if (!$token->checkTokenForAvailability($request->bearerToken())){
+            return response()->json(['message' => 'Unauthorized '], 401 );
+        }
+
         $user = User::find($user_id);
-        $user_temp = $user;
         $result = $user->delete();
         if ($result){
-            return response()->json(['message' => 'User has been deleted','data' => $user_temp],201 );
+            return response()->json(['message' => 'User has been deleted'],200 );
         }
         else
         {
@@ -86,4 +110,8 @@ class UserAPI extends Controller
         }
 
     }
+
+
+
+
 }
